+++ index.html
@@ -28,6 +28,8 @@
   }
 </style>
 
-<script src="script.js"></script>
+<script type="module" src="script.js"></script>
+<script type="module" src="api.js"></script>
+<script type="module" src="render.js"></script>
 
 </html>
